v0.0.164

+++ src/sys/wasm.rs
@@ -14,11 +14,11 @@ use core::ptr;
 pub(crate) unsafe fn map(size: usize) -> *mut u8 {
     debug_assert!(size % crate::page::PAGE_SIZE == 0);
     let pages = size / crate::page::PAGE_SIZE;
-    let old = wasm32::memory_grow(0, pages as u32);
-    if old == usize::MAX as u32 {
+    let old = wasm32::memory_grow(0, pages);
+    if old == usize::MAX {
         return ptr::null_mut();
     }
-    (old as usize * crate::page::PAGE_SIZE) as *mut u8
+    (old * crate::page::PAGE_SIZE) as *mut u8
 }
 
 /// WASM cannot release linear-memory pages; kept mapped by design.
